http: Move query string parsing on request class
Parsing the path?a=b&c=d strings is now method(s) of httpd::connection,
but naturally belong to class request. Move them there. Soon there will
appear the mirroring encoding methods.
